Moving BookCard into a separate component, fixing tailwing css errors, leaving comments
